api/requirements.txt
bentoml==1.4.19
catboost==1.2.8
joblib==1.4.2
pandas==2.3.1
pymysql==1.1.1
python-dotenv==1.1.1
scikit-learn==1.7.1
SQLAlchemy==1.4.54
aiomysql==0.2.0